Add showing tasks list for parents

+++ app/src/main/java/com/paychi/dima/paychi/models/TaskItem.java
@@ -1,11 +1,11 @@
 package com.paychi.dima.paychi.models;
 
 public class TaskItem {
-    private  int itemId;
+    private long itemId;
     private String title;
     private String description;
     private String cost;
-    private int listId;
+    private long listId;
     private int state;
     private int target;
 }
